requirements.txt
# ─────────────────────────────────────────────────────────────────────────────
# Disease Prediction Modeling Pipeline — Requirements
# Python >= 3.9 recommended
# ─────────────────────────────────────────────────────────────────────────────

# Core scientific stack
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.11.0

# Machine Learning — classical
scikit-learn>=1.3.0
xgboost>=2.0.0
lightgbm>=4.0.0
catboost>=1.2.0
imbalanced-learn>=0.11.0       # SMOTE, ADASYN, etc.

# Machine Learning — deep learning
torch>=2.1.0
pytorch-tabnet>=4.1.0          # TabNet

# Hyperparameter optimisation
optuna>=3.4.0
scikit-optimize>=0.9.0

# Interpretability
shap>=0.44.0
lime>=0.2.0.1

# Visualisation
matplotlib>=3.7.0
seaborn>=0.13.0
plotly>=5.17.0

# Statistical testing
statsmodels>=0.14.0
pingouin>=0.5.4                # Kruskal-Wallis, chi-square wrappers

# Data quality & utilities
pyarrow>=14.0.0                # fast parquet I/O
pyyaml>=6.0.1
tqdm>=4.66.0
joblib>=1.3.0
loguru>=0.7.2

# Calibration
netcal>=1.3.5                  # reliability diagrams, ECE

# Notebooks
jupyterlab>=4.0.0
ipywidgets>=8.1.0
